Move the credential component to the common components folder.
